generate dynamic topic from json payload on publish
